+++ newlib/libm/common/sinf.c
@@ -40,7 +40,7 @@
 float
 sinf (float y)
 {
-  double x = y;
+  double x = (double) y;
   double s;
   int n;
   const sincos_t *p = &__sincosf_table[0];
